Remove handle_handshake and accept from Service's API
